# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Upload Auto Service REST API",
        "contact": {
            "name": "Kirill Danilov",
            "email": "danilov@atlasbiomed.com"
        }
    },
    "paths": {
        "/api/v1/product/{userID}": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get product code request",
                "operationId": "getProductCode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to get product code for",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modeldto.ResponseProductCode"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "417": {
                        "description": "Expectation Failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/status/{userID}": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get processing status request",
                "operationId": "getProcessingStatus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to get status for",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modeldto.ResponseProcessingStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "417": {
                        "description": "Expectation Failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "modeldto.ResponseProcessingStatus": {
            "type": "object",
            "properties": {
                "current_status": {
                    "type": "string",
                    "example": "done"
                }
            }
        },
        "modeldto.ResponseProductCode": {
            "type": "object",
            "properties": {
                "product_code": {
                    "type": "string",
                    "example": "upload_23andme_v5_b2c_array_txt"
                }
            }
        }
    }
}